WorkCardコンポーネントのProps型をWorksPropsに変更し、ダッシュボードページで制作実績リストをDashboardWorksListコンポーネントに置き換えました。
