+++ src/types/WorksProps.ts
@@ -1,0 +1,7 @@
+export type WorksProps = {
+  slug: string;
+  imageUrl: string | null;
+  title: string;
+  url: string | null;
+  description: string | null;
+};
